gradingapp/grading-agent/requirements.txt
fastapi==0.115.6
uvicorn==0.32.1
anthropic==0.40.0
python-dotenv==1.0.1
